Add files via upload
Fixed it so the right joints are being changed by the sinusodial signal. Added it so the program attempts to update an unseen joints position according to its momentum when it was last seen. Not very good when joint is moving slowly as it gets slowly covered meaning its momentum is tiny by the time it is fully covered.